% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.lag.partial.R
\name{corr.lag.partial}
\alias{corr.lag.partial}
\title{Partial Cross-correlation function}
\usage{
corr.lag.partial(matH, k)
}
\arguments{
\item{matH}{A matrix with multiple columns (more than 3 columns)}

\item{k}{The lag order (integer)}
}
\value{
Partial corss-correlation at k lags and the correlation statistics at k lags.
}
\description{
A function used to obtain partial cross-correlation function for a give lag order
}
\details{
This function obtains the partial corss-correlation and the simple correlation.
To obtain the partial cross-correlation, this function uses the first column of
the input matrix and k-lagged values of the rest of the matrix.
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
