% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.lag.R
\name{corr.lag}
\alias{corr.lag}
\title{Correlation Function}
\usage{
corr.lag(matH, k)
}
\arguments{
\item{matH}{The matrix with the column size of 2}

\item{k}{The lag order (integer)}
}
\value{
Correlation
}
\description{
The correlation statistics for a given lag order
}
\details{
The function obtains the simple correlation statistics. The values in the first column of input matrix
is interacted with the k-lagged values in the second column.
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
