% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossqreg.partial.R
\name{crossqreg.partial}
\alias{crossqreg.partial}
\title{Paritial Cross-Quantilogram}
\usage{
crossqreg.partial(DATA1, DATA2, vecA, k)
}
\arguments{
\item{DATA1}{An input matrix (T x p1)}

\item{DATA2}{An input matrix (T x p2)}

\item{vecA}{A vector of probability values at which sample quantiles are estiamted}

\item{k}{The lag order}
}
\value{
The partial corss-quantilogram and the cross-quantilogram
}
\description{
Returns the partial cross-quantilogram
}
\details{
This function obtains the partial corss-quantilogram and the cross-quantilogram.
To obtain the partial cross-correlation given an input matrix, this function interacts
the values of the first column and the k-lagged values of the rest of the matrix.
}
\references{
Han, H., Linton, O., Oka, T., and Whang, Y. J. (2016).
"The cross-quantilogram: Measuring quantile dependence and testing directional predictability between time series."
\emph{Journal of Econometrics}, 193(1), 251-270.
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
