% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.hit.R
\name{q.hit}
\alias{q.hit}
\title{Quantile Hit}
\usage{
q.hit(DATA, vecA)
}
\arguments{
\item{DATA}{A matrix that has time-series observations in its columns}

\item{vecA}{A vector of probabilty values at which sample quantiles are estimated}
}
\value{
\preformatted{ A matrix of quantile-hits
}
}
\description{
Returns the matrix of quantil-hits
}
\details{
This function generates the quantile hits given a vector of probabilty values.
The quantile hits are obtained for each column of an input matrix.
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
