% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreg.hit.R
\name{qreg.hit}
\alias{qreg.hit}
\title{Quantile Hit}
\usage{
qreg.hit(DATA1, DATA2, vecA)
}
\arguments{
\item{DATA1}{An input matrix (T x p1+1) with the first column of the dependent varaible and the the rest of columns with regressors}

\item{DATA2}{An input matrix (T x p2+1) with the first column of the dependent varaible and the the rest of columns with regressors}

\item{vecA}{A vector of probabilty values at which sample quantiles are estimated}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A matrix of quantile-hits
}\if{html}{\out{</div>}}
}
\description{
Returns the matrix of quantil-hits
}
\details{
This function generates the quantile hits based on quantile regression,
given a vector of probabilty values. The quantile regressions are esimated
for each matrix of data and a pair of quantile hits are produced.
}
\references{
Koenker, R., and Bassett Jr, G. (1978).
"Regression quantiles." Econometrica, 46(1), 33-50.
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
