% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilogram-package.R
\docType{package}
\name{quantilogram-package}
\alias{quantilogram}
\alias{quantilogram-package}
\title{quantilogram: Cross-Quantilogram}
\description{
Estimation and inference methods for the cross-quantilogram. The cross-quantilogram is a measure of nonlinear dependence between two variables, based on either unconditional or conditional quantile functions. The cross-quantilogram can be considered as an extension of the correlogram, which is a correlation function over multiple lag periods and mainly focuses on linear dependency. One can use the cross-quantilogram to detect the presence of directional predictability from one time series to another. This package provides a statistical inference method based on the stationary bootstrap. See Linton and Whang (2007) <doi:10.1016/j.jeconom.2007.01.004> for univariate time series analysis and Han, Linton, Oka and Whang (2016) <doi:10.1016/j.jeconom.2016.03.001> for multivariate time series analysis.
}
\author{
\strong{Maintainer}: Tatsushi Oka \email{oka.econ@gmail.com}

Other contributors:
\itemize{
  \item Heejon Han [contributor]
  \item Oliver Linton [contributor]
  \item Yoon-Jae Whang [contributor]
}

}
\keyword{internal}
