% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.index.R
\name{sb.index}
\alias{sb.index}
\title{Stationary Bootstrap Index}
\usage{
sb.index(Nsize, gamma)
}
\arguments{
\item{Nsize}{The size of the stationary bootstrap resample}

\item{gamma}{A parameter for the stationary boostrap.}
}
\value{
A vector of indicies for the stationary bootstrap
}
\description{
A subfunction for the statioanry bootstrap
}
\details{
This function resamples blocks of indicies with random block lengths.
This code follows the MATLAB file of the Oxford MFE Toolbox written by Kevin Sheppard.
}
\references{
The Oxford MFE toolbox (http://www.kevinsheppard.com/wiki/MFE_Toolbox) by Kevin Sheppard
}
\author{
Heejoon Han, Oliver Linton, Tatsushi Oka and Yoon-Jae Whang
}
